Adding a whole bunch of stuff around login functionality.  I typically would not recommend adding this much in one commit, however it is mostly all around login functionality.  There is a IMemberRepository, which is currently implemented by FakeMemberRepository (which currently has three users, timbarcz, chrissutton, and chrismissal - all passwords are set to password).
I've added in flash messages to show error and informational messages.  This should be styled at some point.

I've also added a userStatus view which is rendered as part of the master page and shows the current user's status.

There's quite a bit else that is coming along with this commit (Authenticator, Cryptography, FormsAuth stuff) so I would recommend posting a question to the discussion group if you have specific questions.