I would normally throat punch someone for putting a commit this large into source control...however since we don't have local commits (*hint *hint) and Chris and I are the main contributors I'll explain to him what I did. We'll lose history anyway if we ever change source control systems (*hint *hint).
All tests pass (both integration and unit)

Ok so what did I change. I made it so we can accept guest registrations.  I spent a lot of time around Meeting and Attendees and the Domain.  I then worked on persistence of a meeting with attendees (of the various types).

More to do...not nearly as close to launching as I'd hoped.